requirements.txt
certifi==2020.6.20
flask==1.1.2
flask-cors
flask-compress
psycopg2-binary